requirements.txt
torch
transformers
accelerate
pyyaml
pillow
numpy
scipy
tqdm
opencv-python
scikit-image
pycocotools
ultralytics
